Deprecate `InputData` interface
- Move to new `InputDataInterface`
- Old interface now simply extends this for compatibility
